Refactor request services into focused components
