+++ src/Pathfinding.App.Console/Injection/Modules.cs
@@ -39,7 +39,10 @@ internal static class Modules
 
         builder.RegisterInstance(new SqliteUnitOfWorkFactory(Settings.Default.ConnectionString)).As<IUnitOfWorkFactory>().SingleInstance();
 
-        builder.RegisterType<RequestService<GraphVertexModel>>().As<IRequestService<GraphVertexModel>>().SingleInstance();
+        builder.RegisterType<GraphRequestService<GraphVertexModel>>().As<IGraphRequestService<GraphVertexModel>>().SingleInstance();
+        builder.RegisterType<RangeRequestService<GraphVertexModel>>().As<IRangeRequestService<GraphVertexModel>>().SingleInstance();
+        builder.RegisterType<StatisticsRequestService>().As<IStatisticsRequestService>().SingleInstance();
+        builder.RegisterType<GraphInfoRequestService>().As<IGraphInfoRequestService>().SingleInstance();
 
         builder.RegisterType<IncludeSourceVertex<GraphVertexModel>>().SingleInstance().WithAttributeFiltering()
             .WithMetadata(MetadataKeys.Order, 2).Keyed<Command>(KeyFilters.IncludeCommands);
